Fix zip example in scala
The previous example for(f <- foo; b <- bar) iterates through all possible
combinations of foo and bar. 
That is (1, a), (1,b), (1,c), (2,a), (2,b), (2,c), (3,a), (3,b), (3,c).
So all values of a, b, c are assigned as 3 (from the last 3 iterations).

Instead use the zip method of bar to create the tuples and add them to the Map.